Move mock data to cypress/fixtures, deleted mock data
