+++ cypress/e2e/admin/admin/Search.cy.js
@@ -1,4 +1,4 @@
-import { adminList } from "../../../../src/data/mock/Admin";
+import adminList from "../../../fixtures/Admin.json";
 
 describe("Admin Search", () => {
   beforeEach(() => {
