+++ cypress/fixtures/Judges.json
@@ -1,0 +1,50 @@
+[
+  {
+    "type": "professor",
+    "name": "p_Allan Knight"
+  },
+  {
+    "type": "student",
+    "name": "s_Menthy Wu"
+  },
+  {
+    "type": "industry",
+    "name": "i_Shing Hung"
+  },
+  {
+    "type": "professor",
+    "name": "p_Neftali"
+  },
+  {
+    "type": "student",
+    "name": "s_Sachin Chopra"
+  },
+  {
+    "type": "industry",
+    "name": "i_Raidah"
+  },
+  {
+    "type": "professor",
+    "name": "p_Mariam"
+  },
+  {
+    "type": "student",
+    "name": "s_Divyank Shah"
+  },
+  {
+    "type": "industry",
+    "name": "i_Sally"
+  },
+  {
+    "type": "professor",
+    "name": "p_Joe"
+  },
+  {
+    "type": "student",
+    "name": "s_John"
+  },
+  {
+    "type": "industry",
+    "name": "i_Kyle"
+  }
+]
