+++ cypress/fixtures/Mentors.json
@@ -1,0 +1,146 @@
+[
+  {
+    "uid": 1,
+    "name": "Raidah Fairooz",
+    "email": "rfair008@ucr.edu",
+    "discord": "Raidah#2067",
+    "status": "online",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 2,
+    "name": "Minsoo Kim",
+    "email": "mkim001@ucr.edu",
+    "discord": "Soup#2023",
+    "status": "online",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 3,
+    "name": "Menthy Wu",
+    "email": "mwu171@ucr.edu",
+    "discord": "_wmx",
+    "status": "onsite",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 4,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "yiyii#1111",
+    "status": "onsite",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 5,
+    "name": "Tiffany Wang",
+    "email": "twang123@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 6,
+    "name": "Tiffany Wu",
+    "email": "twu7680@ucr.edu",
+    "discord": "yuung#8376",
+    "status": "online",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 7,
+    "name": "Celine Don",
+    "email": "cdon@hotmai.com",
+    "discord": "duuknh#0298",
+    "status": "online",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 8,
+    "name": "AJ Cena",
+    "email": "ajmm@gmail.com",
+    "discord": "zzz_11#2067",
+    "status": "onsite",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 9,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "onsite",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 10,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 11,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 12,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 13,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 14,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 15,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "online",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 16,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  }
+]
