+++ cypress/fixtures/PieCharts.json
@@ -1,0 +1,162 @@
+{
+  "Status": {
+    "size": 4,
+    "data": [
+      {
+        "id": "rejected",
+        "label": "rejected",
+        "value": 34
+      },
+      {
+        "id": "accepted",
+        "label": "accepted",
+        "value": 56
+      },
+      {
+        "id": "pending",
+        "label": "pending",
+        "value": 56
+      }
+    ]
+  },
+  "Majors": {
+    "size": 4,
+    "data": [
+      {
+        "id": "rejected",
+        "label": "rejected",
+        "value": 34
+      },
+      {
+        "id": "accepted",
+        "label": "accepted",
+        "value": 56
+      },
+      {
+        "id": "pending",
+        "label": "pending",
+        "value": 56
+      }
+    ]
+  },
+  "Age": {
+    "size": 4,
+    "data": [
+      {
+        "id": "rejected",
+        "label": "rejected",
+        "value": 34
+      },
+      {
+        "id": "accepted",
+        "label": "accepted",
+        "value": 56
+      },
+      {
+        "id": "pending",
+        "label": "pending",
+        "value": 56
+      }
+    ]
+  },
+  "School": {
+    "size": 6,
+    "data": [
+      {
+        "id": "rejected",
+        "label": "rejected",
+        "value": 34
+      },
+      {
+        "id": "accepted",
+        "label": "accepted",
+        "value": 56
+      },
+      {
+        "id": "pending",
+        "label": "pending",
+        "value": 56
+      }
+    ]
+  },
+  "Grade": {
+    "size": 6,
+    "data": [
+      {
+        "id": "rejected",
+        "label": "rejected",
+        "value": 34
+      },
+      {
+        "id": "accepted",
+        "label": "accepted",
+        "value": 56
+      },
+      {
+        "id": "pending",
+        "label": "pending",
+        "value": 56
+      }
+    ]
+  },
+  "Gender": {
+    "size": 4,
+    "data": [
+      {
+        "id": "rejected",
+        "label": "rejected",
+        "value": 34
+      },
+      {
+        "id": "accepted",
+        "label": "accepted",
+        "value": 56
+      },
+      {
+        "id": "pending",
+        "label": "pending",
+        "value": 56
+      }
+    ]
+  },
+  "Shirt Size": {
+    "size": 4,
+    "data": [
+      {
+        "id": "rejected",
+        "label": "rejected",
+        "value": 34
+      },
+      {
+        "id": "accepted",
+        "label": "accepted",
+        "value": 56
+      },
+      {
+        "id": "pending",
+        "label": "pending",
+        "value": 56
+      }
+    ]
+  },
+  "Dietary Restrictions": {
+    "size": 4,
+    "data": [
+      {
+        "id": "rejected",
+        "label": "rejected",
+        "value": 34
+      },
+      {
+        "id": "accepted",
+        "label": "accepted",
+        "value": 56
+      },
+      {
+        "id": "pending",
+        "label": "pending",
+        "value": 56
+      }
+    ]
+  }
+}
