+++ cypress/fixtures/TeamsCopy.json
@@ -1,0 +1,162 @@
+[
+  {
+    "hidden": false,
+    "name": "Citrus hack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Rosehack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Bearhack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Design",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Citrus hack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Rosehack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Bearhack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Design",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Citrus hack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Rosehack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Bearhack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Design",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Citrus hack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Rosehack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Bearhack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Design",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Citrus hack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Rosehack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Bearhack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Design",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Citrus hack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Rosehack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Bearhack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Design",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Citrus hack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Rosehack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Bearhack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Design",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Citrus hack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Rosehack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Bearhack",
+    "rounds": []
+  },
+  {
+    "hidden": false,
+    "name": "Design",
+    "rounds": []
+  }
+]
