+++ cypress/fixtures/Volunteers.json
@@ -1,0 +1,146 @@
+[
+  {
+    "uid": 1,
+    "name": "Raidah Fairooz",
+    "email": "rfair008@ucr.edu",
+    "discord": "Raidah#2067",
+    "status": "pending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 2,
+    "name": "Minsoo Kim",
+    "email": "mkim001@ucr.edu",
+    "discord": "Soup#2023",
+    "status": "pending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 3,
+    "name": "Menthy Divyank",
+    "email": "couple002@ucr.edu",
+    "discord": "disney#2033",
+    "status": "confirm",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 4,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "confirm",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 5,
+    "name": "John Cena",
+    "email": "digdag@gmail.com",
+    "discord": "dorm#3333",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 6,
+    "name": "John Cena",
+    "email": "john666@hotmail.com",
+    "discord": "kickkk#8906",
+    "status": "confirm",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 7,
+    "name": "John Cena",
+    "email": "cesss2938@ucr.edu",
+    "discord": "bnxgsw#000000",
+    "status": "confirm",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 8,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "confirm",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 9,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "confirm",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 10,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "confirm",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 11,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 12,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 13,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 14,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "pending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 15,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  },
+  {
+    "uid": 16,
+    "name": "John Cena",
+    "email": "cantseeme@ucr.edu",
+    "discord": "babydonthurtme#2067",
+    "status": "not attending",
+    "selected": false,
+    "hidden": false
+  }
+]
