+++ src/components/Admin/Services/Statistics.jsx
@@ -3,7 +3,7 @@ import PieChart from "@/components/Admin/Services/PieChart";
 import Tag from "@/components/Admin/Tag";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
-import { charts } from "@/data/mock/PieCharts";
+import charts from "../../../../cypress/fixtures/PieCharts.json";
 const Statistics = () => {
   return (
     <div className="h-full font-poppins flex flex-col py-4 gap-3">
