--- src/data/mock/Admin.js
@@ -1,135 +1,0 @@
-export const adminList = [
-  {
-    uid: 1,
-    name: "Dewi Norton",
-    position: "lead",
-    status: "pending",
-    affiliation: "recruitment",
-    email: "dewinorton@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 2,
-    name: "Amie Nguyen",
-    position: "lead",
-    status: "reject",
-    affiliation: "director",
-    email: "amienguyenn@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 3,
-    name: "Larry Larsen",
-    position: "",
-    status: "reject",
-    affiliation: "marketing",
-    email: "larrylarsen@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 4,
-    name: "Big Chungus",
-    position: "",
-    status: "accept",
-    affiliation: "operations",
-    email: "bigchungus101@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 5,
-    name: "Mario Kart",
-    position: "",
-    status: "accept",
-    affiliation: "sponsorship",
-    email: "mariomoviegoated101@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 6,
-    name: "Ash Ketchum",
-    position: "",
-    status: "pending",
-    affiliation: "software",
-    email: "ash.ketchum12@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 7,
-    name: "Tiffany Wang",
-    position: "",
-    affiliation: "UIUX",
-    status: "accept",
-    email: "twang777@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 8,
-    name: "Benjamin Wang",
-    affiliation: "sponsorship",
-    status: "pending",
-    email: "benjjj@ucr.edu",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 9,
-    name: "Catty Sun",
-    affiliation: "operations",
-    status: "accept",
-    email: "ash.catty1278@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 10,
-    name: "Cindy Koy",
-    affiliation: "industry",
-    status: "accept",
-    email: "kcindy@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 11,
-    name: "Venessal Teng",
-    affiliation: "industry",
-    status: "accept",
-    email: "vteng666@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 12,
-    name: "Savv Kun",
-    affiliation: "marketing",
-    status: "reject",
-    email: "msuusse8966@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 13,
-    name: "Alex Jenkins",
-    affiliation: "industry",
-    status: "pending",
-    email: "akkkk74@email.com",
-    selected: false,
-    hidden: false,
-  },
-  {
-    uid: 14,
-    name: "Astin Downy",
-    affiliation: "industry",
-    status: "pending",
-    email: "Lqy17837@hotmail.com",
-    selected: false,
-    hidden: false,
-  },
-];
