--- src/data/mock/CheckInUser.js
@@ -1,20 +1,0 @@
-export const checkInUser = {
-  user1: {
-    name: "Menthyt Wu",
-    email: "mwu171@ucr.edu",
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdqnA8Mv1vcnqypi3YepWr3c7bP97wd6CJuDaoz1N5m=s576-c-no-mo",
-    Lunch: true,
-    Dinner: true,
-  },
-  user2: {
-    name: "Divyank Shah",
-    email: "dshah048@ucr.edu",
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-    "Web Dev Workshop": true,
-    "Participents Check In": true,
-    "Career Workshop": true,
-    "Parjama Party": true,
-  },
-};
