--- src/data/mock/Events.js
@@ -1,32 +1,0 @@
-export const mockEvents = [
-  { name: "Participents Check In", hidden: false },
-  { name: "Web Dev Workshop", hidden: false },
-  { name: "Unity Workshop", hidden: false },
-  { name: "Lunch", hidden: false },
-  { name: "Career Workshop", hidden: false },
-  { name: "Data Science Workshop", hidden: false },
-  { name: "Dinner", hidden: false },
-  { name: "Boba Social", hidden: false },
-  { name: "Parjama Party", hidden: false },
-  { name: "Breakfast", hidden: false },
-  { name: "Participents Check In", hidden: false },
-  { name: "Web Dev Workshop", hidden: false },
-  { name: "Unity Workshop", hidden: false },
-  { name: "Lunch", hidden: false },
-  { name: "Career Workshop", hidden: false },
-  { name: "Data Science Workshop", hidden: false },
-  { name: "Dinner", hidden: false },
-  { name: "Boba Social", hidden: false },
-  { name: "Parjama Party", hidden: false },
-  { name: "Breakfast", hidden: false },
-  { name: "Participents Check In", hidden: false },
-  { name: "Web Dev Workshop", hidden: false },
-  { name: "Unity Workshop", hidden: false },
-  { name: "Lunch", hidden: false },
-  { name: "Career Workshop", hidden: false },
-  { name: "Data Science Workshop", hidden: false },
-  { name: "Dinner", hidden: false },
-  { name: "Boba Social", hidden: false },
-  { name: "Parjama Party", hidden: false },
-  { name: "Breakfast", hidden: false },
-];
