--- src/data/mock/Judges.js
@@ -1,222 +1,0 @@
-export const judgeList = [
-  {
-    uid: 1,
-    name: "Dewi Norton",
-    status: "pending",
-    type: "professor",
-    email: "dewinorton@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 2,
-    name: "Amie Nguyen",
-    status: "not attending",
-    type: "professor",
-    email: "amienguyenn@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 3,
-    name: "Larry Larsen",
-    status: "not attending",
-    type: "professor",
-    email: "larrylarsen@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 4,
-    name: "Big Chungus",
-    status: "confirm",
-    type: "professor",
-    email: "bigchungus101@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 5,
-    name: "Mario Kart",
-    status: "confirm",
-    type: "student",
-    email: "mariomoviegoated101@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 6,
-    name: "Ash Ketchum",
-    type: "industry",
-    status: "pending",
-    email: "ash.ketchum12@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 7,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 8,
-    name: "Amy Cheng",
-    type: "industry",
-    status: "confirm",
-    email: "chenamy490@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 9,
-    name: "Kim Alexander",
-    type: "industry",
-    status: "not attending",
-    email: "kim5867@hotmai.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 10,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 11,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 12,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 13,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 14,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 15,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 16,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 17,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 18,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 19,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-  {
-    uid: 20,
-    name: "Tiffany",
-    type: "student",
-    status: "pending",
-    email: "twang123@email.com",
-    selected: false,
-    hidden: false,
-    image:
-      "https://lh3.googleusercontent.com/a/AAcHTtdDqygBXOmrnoQfN4LzUzTEI0g5y88-dO6jqCMSmJHvTQ=s96-c",
-  },
-];
