--- src/data/mock/Judging.js
@@ -1,213 +1,0 @@
-export const teamsCopy = [
-  {
-    hidden: false,
-    name: "Citrus hack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Rosehack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Bearhack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Design",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Citrus hack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Rosehack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Bearhack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Design",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Citrus hack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Rosehack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Bearhack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Design",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Citrus hack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Rosehack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Bearhack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Design",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Citrus hack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Rosehack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Bearhack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Design",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Citrus hack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Rosehack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Bearhack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Design",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Citrus hack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Rosehack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Bearhack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Design",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Citrus hack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Rosehack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Bearhack",
-    rounds: [],
-  },
-  {
-    hidden: false,
-    name: "Design",
-    rounds: [],
-  },
-];
-
-export const judges = [
-  {
-    type: "professor",
-    name: "p_Allan Knight",
-  },
-  {
-    type: "student",
-    name: "s_Menthy Wu",
-  },
-  {
-    type: "industry",
-    name: "i_Shing Hung",
-  },
-  {
-    type: "professor",
-    name: "p_Neftali",
-  },
-  {
-    type: "student",
-    name: "s_Sachin Chopra",
-  },
-  {
-    type: "industry",
-    name: "i_Raidah",
-  },
-  {
-    type: "professor",
-    name: "p_Mariam",
-  },
-  {
-    type: "student",
-    name: "s_Divyank Shah",
-  },
-  {
-    type: "industry",
-    name: "i_Sally",
-  },
-  {
-    type: "professor",
-    name: "p_Joe",
-  },
-  {
-    type: "student",
-    name: "s_John",
-  },
-  {
-    type: "industry",
-    name: "i_Kyle",
-  },
-];
