--- src/data/mock/Participants.js
@@ -1,385 +1,0 @@
-export const participantList = [
-  {
-    uid: "1",
-    name: "Menthy Wu",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "18",
-    school: "University of California - Riverside",
-    grade: "junior",
-    gender: "Female",
-    size: "XS",
-    restriction: ["Vegan", "Cheese"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "The team",
-    major: "Computer Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "2",
-    name: "Divyank Shah",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "20",
-    school: "University of California - Riverside",
-    grade: "junior",
-    gender: "Male",
-    size: "S",
-    restriction: "N/A",
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "The deer",
-    major: "Computer Science with Business Applications",
-    status: "accept",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "3",
-    name: "Shing Hung",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "The team",
-    major: "Data Science",
-    status: "reject",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "4",
-    name: "Sachin Chopra",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "The deer",
-    major: "Environmental Engineering",
-    status: "accept",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-
-  {
-    uid: "5",
-    name: "Tiffany Wang",
-    email: "xxx022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Chemical Engineering",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-
-  {
-    uid: "6",
-    name: "Alex Wu",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Engineering",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-
-  {
-    uid: "7",
-    name: "Cindy Jones",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "8",
-    name: "AJ Hall",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Mechanical Engineering",
-    status: "accept",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "9",
-    name: "Lucy Pan",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Environmental Engineering",
-    status: "reject",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "10",
-    name: "Ben Du",
-    email: "yhung987@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "11",
-    name: "Wendy Mu",
-    email: "yhung0465@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "12",
-    name: "Monique Hun",
-    email: "yhung012@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Science with Business Applications",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "13",
-    name: "Eddie Shappun",
-    email: "mzhun022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "14",
-    name: "Nancy Zun",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "15",
-    name: "Jack xx",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Electrical Engineering",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "16",
-    name: "Bellay Chappan",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Engineering",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "17",
-    name: "Cathy Shannon",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "18",
-    name: "Menthy Wu",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Mechanical Engineering",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "19",
-    name: "Menthy Wu",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Data Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-  {
-    uid: "20",
-    name: "Menthy Wu",
-    email: "yhung022@ucr.edu",
-    phone: "000-000-0000",
-    age: "22",
-    school: "University of California - Irvine",
-    grade: "junior",
-    gender: "Other",
-    size: "XS",
-    restriction: ["Vegan", "No Nuts"],
-    resume: "https://shahdivyank.github.io/Divyank_Shah_Resume.pdf",
-    team: "Citrus",
-    major: "Computer Science",
-    status: "pending",
-    selected: false,
-    hidden: false,
-    dropdown: true,
-  },
-];
