--- src/data/mock/Teams.js
@@ -1,199 +1,0 @@
-export const teamList = [
-  {
-    uid: 1,
-    name: "The team",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "winner",
-    selected: false,
-    members: ["Menthy Wu", "Mengxuan Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 2,
-    name: "The deer",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "pending",
-    selected: false,
-    members: ["Menthy Wu", "Mengxuan Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 3,
-    name: "The team",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "qualify",
-    selected: false,
-    members: ["Menthy Wu", "Mengxuan Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 4,
-    name: "The deer",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "disqualify",
-    selected: false,
-    members: ["Menthy Wu", "Mengxuan Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-
-  {
-    uid: 5,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "winner",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 6,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "winner",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["twang022@ucr.edu", "twu278@ucr.edu"],
-    hidden: false,
-  },
-
-  {
-    uid: 7,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "winner",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 8,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "winner",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 9,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "disqualify",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: true,
-  },
-  {
-    uid: 9,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "disqualify",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 9,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "pending",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 9,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "disqualify",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 9,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "disqualify",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 9,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "disqualify",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-  {
-    uid: 9,
-    name: "Citrus",
-    links: [
-      { name: "github", link: "https://github.com" },
-      { name: "devpost", link: "https://rose-hack-2021.devpost.com" },
-    ],
-    status: "disqualify",
-    selected: false,
-    members: ["Tiffany Wang", "Tiffany Wu"],
-    emails: ["yhung022@ucr.edu", "yhung022@ucr.edu"],
-    hidden: false,
-  },
-];
